Tests/LibWeb/Text/expected/display_list/caret-invalidation-in-relative-inline.txt

caret painted: true
caret moved: true
